Refactor curators pages to use Supabase DB, update copy from deals to projects, and remove all mock data
