+++ app/curators/[id]/loading.tsx
