+++ hooks/use-wallet-balance.ts
@@ -1,4 +1,4 @@
-import { useState, useEffect } from "react";
+import { useState, useEffect, useCallback } from "react";
 import { useWallets } from "@privy-io/react-auth";
 import { ethers } from "ethers";
 
@@ -14,30 +14,31 @@ export function useWalletBalance() {
   const [balance, setBalance] = useState<string>("0");
   const [loading, setLoading] = useState(false);
 
-  useEffect(() => {
-    const fetchBalance = async () => {
-      if (!embeddedWallet?.address) return;
-      setLoading(true);
-      try {
-        const BASE_RPC_URL = `https://base-mainnet.infura.io/v3/${process.env.NEXT_PUBLIC_INFURA_PROJECT_ID}`;
-        const provider = new ethers.JsonRpcProvider(BASE_RPC_URL);
-        const usdcContract = new ethers.Contract(USDC_CONTRACT_ADDRESS, USDC_ABI, provider);
-        const [rawBalance, tokenDecimals] = await Promise.all([
-          usdcContract.balanceOf(embeddedWallet.address),
-          usdcContract.decimals(),
-        ]);
-        const balanceInWei = rawBalance.toString();
-        const decimals = Number(tokenDecimals);
-        const formattedBalance = (Number(balanceInWei) / Math.pow(10, decimals)).toFixed(decimals);
-        setBalance(formattedBalance);
-      } catch (error) {
-        setBalance("0");
-      } finally {
-        setLoading(false);
-      }
-    };
-    fetchBalance();
+  const fetchBalance = useCallback(async () => {
+    if (!embeddedWallet?.address) return;
+    setLoading(true);
+    try {
+      const BASE_RPC_URL = `https://base-mainnet.infura.io/v3/${process.env.NEXT_PUBLIC_INFURA_PROJECT_ID}`;
+      const provider = new ethers.JsonRpcProvider(BASE_RPC_URL);
+      const usdcContract = new ethers.Contract(USDC_CONTRACT_ADDRESS, USDC_ABI, provider);
+      const [rawBalance, tokenDecimals] = await Promise.all([
+        usdcContract.balanceOf(embeddedWallet.address),
+        usdcContract.decimals(),
+      ]);
+      const balanceInWei = rawBalance.toString();
+      const decimals = Number(tokenDecimals);
+      const formattedBalance = (Number(balanceInWei) / Math.pow(10, decimals)).toFixed(decimals);
+      setBalance(formattedBalance);
+    } catch (error) {
+      setBalance("0");
+    } finally {
+      setLoading(false);
+    }
   }, [embeddedWallet?.address]);
 
-  return { balance, loading };
+  useEffect(() => {
+    fetchBalance();
+  }, [fetchBalance]);
+
+  return { balance, loading, refetch: fetchBalance };
 }
